+++ rust/src/lib.rs
@@ -39,7 +39,8 @@ pub mod prelude {
     pub use crate::contract_interface::wasm_interface::ContractInterfaceResult;
     pub use crate::contract_interface::*;
     pub use crate::delegate_host::{
-        decode_secret_key_list, encode_secret_key_list, error_codes, DelegateCtx,
+        decode_contract_id_list, decode_secret_key_list, encode_contract_id_list,
+        encode_secret_key_list, error_codes, DelegateCtx,
     };
     pub use crate::delegate_interface::wasm_interface::DelegateInterfaceResult;
     pub use crate::delegate_interface::*;